+++ CHANGES.rst
@@ -3,7 +3,8 @@ Changes
 
 Unreleased
 ----------
-- Add ability to specify alternative user_model for tracking
+- Add ability to specify alternative user_model for tracking (gh-371)
+- Add util function ``bulk_create_with_history`` to allow bulk_create with history saved
 
 2.1.1 (2018-06-15)
 ------------------